live updates for tile layer changes
