+++ react/src/pages/MapProject/MapProject.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState, useEffect, useMemo } from 'react';
 import { useLocation, useParams } from 'react-router-dom';
 import { useQueryClient } from '@tanstack/react-query';
 import { Layout, Flex } from 'antd';
@@ -25,6 +25,31 @@ import HeaderNavBar from '@hazmapper/components/HeaderNavBar';
 import styles from './MapProject.module.css';
 import { Spinner } from '@hazmapper/common_components';
 import { Panel as BasePanel } from '@hazmapper/components/Panel';
+import { zodResolver } from '@hookform/resolvers/zod';
+import * as z from 'zod';
+import { useForm, FormProvider } from 'react-hook-form';
+
+export const tileLayerSchema = z.object({
+  id: z.number(),
+  name: z.string().min(1, 'Required'),
+  type: z.string(),
+  url: z.string().url().min(1, 'Required'),
+  attribution: z.string(),
+  tileOptions: z.object({
+    maxZoom: z.number().nullish(),
+    minZoom: z.number().nullish(),
+    maxNativeZoom: z.number().nullish(),
+    format: z.string().nullish(),
+    layers: z.string().nullish(),
+  }),
+  uiOptions: z.object({
+    zIndex: z.number(),
+    opacity: z.number(),
+    isActive: z.boolean(),
+    showInput: z.boolean().nullish(),
+    showDescription: z.boolean().nullish(),
+  }),
+});
 
 interface MapProjectProps {
   /**
@@ -180,86 +205,113 @@ const LoadedMapProject: React.FC<LoadedMapProject> = ({
 
   const { Header, Content, Sider } = Layout;
 
+  const formSchema = z.object({
+    tileLayers: z.array(
+      z.object({
+        layer: tileLayerSchema, // Need to nest tile layer here since useFieldArray will add it's own `id` field, overwriting our own
+      })
+    ),
+  });
+
+  const initialValues = useMemo(
+    () => ({
+      tileLayers: tileServerLayers
+        ?.sort((a, b) => b.uiOptions.zIndex - a.uiOptions.zIndex)
+        .map((layer) => ({ layer })),
+    }),
+    [tileServerLayers]
+  );
+
+  const methods = useForm({
+    defaultValues: initialValues,
+    resolver: zodResolver(formSchema),
+    mode: 'onChange',
+  });
+
+  const { reset } = methods;
+
+  useEffect(() => {
+    reset(initialValues);
+  }, [initialValues, reset]);
+
   return (
-    <Layout style={{ height: '100vh' }}>
-      <Header>
-        <HeaderNavBar />
-        <div className={styles.mapControlBar}>
-          MapTopControlBar TODO https://tacc-main.atlassian.net/browse/WG-260
-        </div>
-      </Header>
-      <Layout>
-        <Sider width="auto">
-          <Flex
-            style={{
-              overflowY: 'auto',
-              height: '100%',
-            }}
-          >
-            <MapProjectNavBar />
-            {activePanel && activePanel !== Panel.Manage && !loading && (
-              <BasePanel
-                panelTitle={activePanel}
-                className={styles.panelContainer}
-              >
-                {activePanel === Panel.Assets && (
-                  <AssetsPanel
-                    project={activeProject}
-                    isPublicView={isPublicView}
-                    featureCollection={featureCollection}
-                  />
-                )}
-                {activePanel === Panel.Filters && (
-                  <Filters
-                    selectedAssetTypes={selectedAssetTypes}
-                    onFiltersChange={setSelectedAssetTypes}
-                    startDate={startDate}
-                    setStartDate={setStartDate}
-                    endDate={endDate}
-                    setEndDate={setEndDate}
-                    toggleDateFilter={toggleDateFilter}
-                    setToggleDateFilter={setToggleDateFilter}
-                  />
-                )}
-                {activePanel === Panel.Layers && (
-                  <LayersPanel
-                    tileLayers={tileServerLayers}
-                    projectId={activeProject.id}
-                    isPublicView={isPublicView}
-                  />
-                )}
-              </BasePanel>
-            )}
-          </Flex>
-        </Sider>
-        <Content>
-          {loading ? (
-            <Spinner />
-          ) : (
-            <>
-              {activePanel === Panel.Manage && (
-                <ManageMapProjectModal isPublicView={isPublicView} />
+    <FormProvider {...methods}>
+      <Layout style={{ height: '100vh' }}>
+        <Header>
+          <HeaderNavBar />
+          <div className={styles.mapControlBar}>
+            MapTopControlBar TODO https://tacc-main.atlassian.net/browse/WG-260
+          </div>
+        </Header>
+        <Layout>
+          <Sider width="auto">
+            <Flex
+              style={{
+                overflowY: 'auto',
+                height: '100%',
+              }}
+            >
+              <MapProjectNavBar />
+              {activePanel && activePanel !== Panel.Manage && !loading && (
+                <BasePanel
+                  panelTitle={activePanel}
+                  className={styles.panelContainer}
+                >
+                  {activePanel === Panel.Assets && (
+                    <AssetsPanel
+                      project={activeProject}
+                      isPublicView={isPublicView}
+                      featureCollection={featureCollection}
+                    />
+                  )}
+                  {activePanel === Panel.Filters && (
+                    <Filters
+                      selectedAssetTypes={selectedAssetTypes}
+                      onFiltersChange={setSelectedAssetTypes}
+                      startDate={startDate}
+                      setStartDate={setStartDate}
+                      endDate={endDate}
+                      setEndDate={setEndDate}
+                      toggleDateFilter={toggleDateFilter}
+                      setToggleDateFilter={setToggleDateFilter}
+                    />
+                  )}
+                  {activePanel === Panel.Layers && (
+                    <LayersPanel
+                      projectId={activeProject.id}
+                      isPublicView={isPublicView}
+                    />
+                  )}
+                </BasePanel>
               )}
-              <div className={styles.map}>
-                <Map
-                  baseLayers={tileServerLayers}
-                  featureCollection={featureCollection}
-                />
-              </div>
-              {selectedFeature && (
-                <div className={styles.detailContainer}>
-                  <AssetDetail
-                    selectedFeature={selectedFeature}
-                    onClose={() => toggleSelectedFeature(selectedFeature.id)}
-                    isPublicView={activeProject.public}
-                  />
+            </Flex>
+          </Sider>
+          <Content>
+            {loading ? (
+              <Spinner />
+            ) : (
+              <>
+                {activePanel === Panel.Manage && (
+                  <ManageMapProjectModal isPublicView={isPublicView} />
+                )}
+                <div className={styles.map}>
+                  <Map featureCollection={featureCollection} />
                 </div>
-              )}
-            </>
-          )}
-        </Content>
+                {selectedFeature && (
+                  <div className={styles.detailContainer}>
+                    <AssetDetail
+                      selectedFeature={selectedFeature}
+                      onClose={() => toggleSelectedFeature(selectedFeature.id)}
+                      isPublicView={activeProject.public}
+                    />
+                  </div>
+                )}
+              </>
+            )}
+          </Content>
+        </Layout>
       </Layout>
-    </Layout>
+    </FormProvider>
   );
 };
