+++ react/src/pages/MapProject/index.ts
@@ -1,1 +1,1 @@
-export { default } from './MapProject';
+export { default, tileLayerSchema } from './MapProject';
